services/crawler/requirements.txt
beautifulsoup4==4.12.3
httpx==0.28.1
lxml==5.3.0
meilisearch==0.35.0
psycopg[binary]==3.2.6
pydantic==2.10.6
pytest==8.3.5
PyYAML==6.0.2
readability-lxml==0.8.1
trafilatura==1.12.2

